requirements.txt
nest_asyncio==1.5.6
python-telegram-bot[job-queue]==20.0
openai==0.27.6
requests==2.28.0
beautifulsoup4==4.11.1
urllib3==1.26.14
python-dotenv==1.0.0  
